backend/requirements.txt
flask
flask-cors
tensorflow
keras
pandas
numpy
scikit-learn
joblib
python-dotenv
psycopg2-binary
